Move eval to Core.hs
